\name{BSSasymp-package}
\alias{BSSasymp-package}
\docType{package}
\title{
Asymptotic Covariance Matrices of Some BSS Mixing and Unmixing Matrix Estimates. 
}
\description{
Description: Functions to compute the asymptotic covariance matrices of mixing and unmixing matrix estimates of the following blind source separation (BSS) methods: symmetric and squared symmetric FastICA, regular and adaptive deflation-based FastICA, FOBI, JADE, AMUSE and deflation-based and symmetric SOBI. Also functions to estimate these covariances based on data are available. 
}
\details{
\tabular{ll}{
Package: \tab BSSasymp\cr
Type: \tab Package\cr
Version: \tab 1.1-1\cr
Date: \tab 2015-08-17\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@jyu.fi>

}
\keyword{package}
