\name{BSSasymp-package}
\alias{BSSasymp-package}
\docType{package}
\title{
Asymptotic Covariance Matrices of Some BSS Mixing and Unmixing Matrix Estimates. 
}
\description{
Description: Functions to compute the asymptotic covariance matrices of mixing and unmixing matrix estimates of the following blind source separation (BSS) methods: symmetric and squared symmetric FastICA, regular and adaptive deflation-based FastICA, FOBI, JADE, AMUSE and deflation-based and symmetric SOBI. Also functions to estimate these covariances based on data are available. 
}
\details{
\tabular{ll}{
Package: \tab BSSasymp\cr
Type: \tab Package\cr
Version: \tab 1.2-2\cr
Date: \tab 2021-12-08\cr
License: \tab GPL (>= 2)\cr
}

}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2017), Blind Source Separation Based on Joint Diagonalization in R: The Packages JADE and BSSasymp, \emph{Journal of Statistical Software}, \bold{76}, 1-31, <doi:10.18637/jss.v076.i02>.}
}

\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.k.nordhausen@jyu.fi>

}
\keyword{package}
