\name{BSSasymp-package}
\alias{BSSasymp-package}
\docType{package}
\title{
Asymptotic Covariance Matrices of Some BSS Mixing and Unmixing Matrix Estimates. 
}
\description{
Description: Functions to compute the asymptotic covariance matrices of mixing and unmixing matrix estimates of the following blind source separation (BSS) methods: symmetric and squared symmetric FastICA, regular and adaptive deflation-based FastICA, FOBI, JADE, AMUSE and deflation-based and symmetric SOBI. Also functions to estimate these covariances based on data are available. For details, see Miettinen et al. (2015) <doi:10.1214/15-STS520>, Miettinen et al. (2016) <doi:10.1111/jtsa.12159>, Miettinen et al. (2017) <doi:10.1016/j.sigpro.2016.08.028>, Miettinen et al. (2017) <doi:10.18637/jss.v076.i02>, and references therein.
}
\details{
\tabular{ll}{
Package: \tab BSSasymp\cr
Type: \tab Package\cr
Version: \tab 1.2-0\cr
Date: \tab 2017-09-11\cr
License: \tab GPL (>= 2)\cr
}

}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2017), Blind Source Separation Based on Joint Diagonalization in R: The Packages JADE and BSSasymp, \emph{Journal of Statistical Software}, \bold{76}, 1-31, <doi:10.18637/jss.v076.i02>.}
}

\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@aalto.fi>

}
\keyword{package}
