\name{BSSasymp-package}
\alias{BSSasymp-package}
\docType{package}
\title{
Asymptotic covariance matrices of some BSS mixing and unmixing matrix estimates. 
}
\description{
Description: The package includes functions to compute the asymptotic covariance matrices of mixing and unmixing matrix estimates of the following blind source separation (BSS) methods: adaptive FastICA, FOBI, JADE, AMUSE and deflation-based and symmetric SOBI. Also functions to estimate these covariances based on data are available. 
}
\details{
\tabular{ll}{
Package: \tab BSSasymp\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2013-11-24\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@jyu.fi>

}
\keyword{package}
